rfclient: Split out headers
